取消订单业务代码开发

+++ sky-server/src/main/java/com/sky/mapper/OrderMapper.java
@@ -56,5 +56,5 @@ public interface OrderMapper {
      * @return
      */
     @Select("select *from orders where id = #{id}")
-    Orders getById(Integer id);
+    Orders getById(Long id);
 }
